xDS router and lb integration test: action
