+++ .github/workflows/agentcore_service_test.yml
@@ -20,6 +20,8 @@ on:
       - '.github/actions/common/plugin-change-check/action.yml'
       - '.github/actions/common/entry/action.yml'
       - '.github/actions/common/exit/action.yml'
+      - 'sermant-plugins/sermant-router/router-common/**'
+      - 'sermant-plugins/sermant-router/spring-router-plugin/**'
 concurrency:
   group: ${{ github.workflow }}-${{ github.event.pull_request.number || github.ref }}-${{ github.head_ref }}
   cancel-in-progress: true
@@ -103,14 +105,13 @@ jobs:
       - name: cache xds service package
         uses: actions/cache@v3
         with:
-          path: sermant-integration-tests/xds-service-test/product/
-          key: ${{ runner.os }}-xds-service-product-${{ github.run_id }}
+          path: sermant-integration-tests/xds-service-test/product/sermant-agent-*/
+          key: ${{ runner.os }}-xds-service-product-agent-${{ github.run_id }}
       - name: package agent
         run: |
           sed -i '/sermant-backend/d' pom.xml
           sed -i '/sermant-injector/d' pom.xml
           mvn package -DskipTests -Ptest --file pom.xml
-          cp -r sermant-integration-tests/xds-service-test/product/spring-client/agent sermant-integration-tests/xds-service-test/product/spring-server/
           mkdir -p sermant-agent-${{ env.sermantVersion }}/agent/pluginPackage/agentcore-test-plugin/plugin
           mkdir -p sermant-agent-${{ env.sermantVersion }}/agent/pluginPackage/dynamic-test-first-plugin/plugin
           mkdir -p sermant-agent-${{ env.sermantVersion }}/agent/pluginPackage/dynamic-test-second-plugin/plugin
@@ -194,3 +195,72 @@ jobs:
         uses: ./.github/actions/common/xds-service
       - name: xds service discovery
         uses: ./.github/actions/scenarios/xds-service/xds-service-discovery/client-envoy
+  test-for-xds-router-lb:
+    name: Test for xds router and lb with router plugin
+    runs-on: ubuntu-latest
+    if: needs.set-execution-conditions.outputs.enableXdsServicAction == 'true'
+    needs: [set-execution-conditions, build-agent-and-cache]
+    strategy:
+      matrix:
+        include:
+          - springBootVersion: "2.0.2.RELEASE"
+            snakeyamlVersion: "1.19"
+            springCloudVersion: "Finchley.RELEASE"
+            httpClientVersion: "4.4"
+            okHttp2Version: "2.2.0"
+            okHttp3Version: "3.5.0"
+            httpAsyncClientVersion: "4.0.1"
+          - springBootVersion: "2.1.0.RELEASE"
+            snakeyamlVersion: "1.23"
+            springCloudVersion: "Greenwich.RELEASE"
+            httpClientVersion: "4.4.1"
+            okHttp2Version: "2.3.0"
+            okHttp3Version: "3.9.1"
+            httpAsyncClientVersion: "4.0.2"
+          - springBootVersion: "2.2.0.RELEASE"
+            snakeyamlVersion: "1.25"
+            springCloudVersion: "Hoxton.RELEASE"
+            httpClientVersion: "4.5"
+            okHttp2Version: "2.4.0"
+            okHttp3Version: "3.12.13"
+            httpAsyncClientVersion: "4.1"
+          - springBootVersion: "2.3.0.RELEASE"
+            snakeyamlVersion: "1.26"
+            springCloudVersion: "Hoxton.RELEASE"
+            httpClientVersion: "4.5.3"
+            okHttp2Version: "2.5.0"
+            okHttp3Version: "3.14.9"
+            httpAsyncClientVersion: "4.1.1"
+          - springBootVersion: "2.4.0"
+            snakeyamlVersion: "1.27"
+            springCloudVersion: "2020.0.0"
+            httpClientVersion: "4.5.7"
+            okHttp2Version: "2.6.0"
+            okHttp3Version: "4.2.2"
+            httpAsyncClientVersion: "4.1.2"
+          - springBootVersion: "2.6.2"
+            snakeyamlVersion: "1.29"
+            springCloudVersion: "2021.0.0"
+            httpClientVersion: "4.5.10"
+            okHttp2Version: "2.7.3"
+            okHttp3Version: "4.7.2"
+            httpAsyncClientVersion: "4.1.3"
+          - springBootVersion: "2.7.17"
+            snakeyamlVersion: "1.30"
+            springCloudVersion: "2021.0.3"
+            httpClientVersion: "4.5.13"
+            okHttp2Version: "2.7.5"
+            okHttp3Version: "4.12.0"
+            httpAsyncClientVersion: "4.1.5"
+      fail-fast: false
+    steps:
+      - uses: actions/checkout@v3
+        with:
+          fetch-depth: 100
+      - name: set java version to environment
+        run: |
+          echo "javaVersion=8" >> $GITHUB_ENV
+      - name: xds common operation
+        uses: ./.github/actions/common/xds-service
+      - name: xds router and lb test
+        uses: ./.github/actions/scenarios/xds-service/xds-router-lb
