+++ sermant-integration-tests/xds-service-test/script/router/router-rule/spring-server-virtureservice.yaml
@@ -39,7 +39,7 @@ spec:
           version:
             exact: base
         uri:
-          exact: /router
+          prefix: /
         ignoreUriCase: false
     route:
     - destination:
