+++ sermant-integration-tests/xds-service-test/spring-server/pom.xml
@@ -20,7 +20,6 @@
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
-            <version>${spring.version}</version>
         </dependency>
         <dependency>
             <groupId>org.springframework.cloud</groupId>
@@ -33,7 +32,7 @@
             <plugin>
                 <groupId>org.springframework.boot</groupId>
                 <artifactId>spring-boot-maven-plugin</artifactId>
-                <version>${spring.version}</version>
+                <version>${springboot.version}</version>
                 <configuration>
                     <mainClass>io.sermant.demo.spring.server.SpringServerApplication</mainClass>
                     <finalName>spring-server</finalName>
