+++ sermant-integration-tests/xds-service-test/xds-service-discovery/pom.xml
@@ -13,7 +13,7 @@
     <packaging>pom</packaging>
 
     <properties>
-        <output.basedir>${pom.basedir}/../product/spring-client</output.basedir>
+        <output.basedir>${pom.basedir}/../product/sermant-agent-${project.version}</output.basedir>
         <package.plugin.name>xds-service-discovery</package.plugin.name>
         <package.plugin.dir>${output.basedir}/agent/pluginPackage</package.plugin.dir>
         <package.plugin.type>undefined</package.plugin.type>
